Joseph's chicken coop
Joseph's chicken coop has room for M chickens (M is a known number, define it as you wish). Each hen Hen has a serial number (between 0 and (1 – M )).
Each hen lays eggs, and each egg is given a unique ID number. In addition, each egg's weight at the time of laying (Weight) and the date it was laid (Date ) are also stored.

You must define the Egg class.
You must propose a data structure that will help Joseph manage the chicken coop.
The following are the operations performed in the chicken coop:
· Initializing the data structure - There is no egg in the coop. The operation will be executed in complexity - O(M).
· Adding an egg - The chicken with serial number i laid an egg, with an ID number, weight - W, and the date of laying - Date. The egg must be added to the structure in complexity - O(1).
· Printing a list of all eggs belonging to hen i. Complexity – O(K), where K- the number of eggs belonging to hen i.
·

The hen numbered i has died and been removed from the coop. All the eggs belonging to it are transferred to hen j. The operation will be performed with complexity O(K), where K - the number of eggs of hen j.

A. Propose a suitable data structure that can enable the implementation of the operations with the given complexity.
Draw the structure. You can attach a picture.
B. Real classes: Egg, Hen, Hen_house and all the functions defined above.
C. Write a test program for Joseph's coop.